Refactor verify client endpoint and move to company module
